refactor: share walk util across parsers
