+++ src/parser/sharedWalk.ts
@@ -1,0 +1,14 @@
+import Parser from "tree-sitter";
+
+export function walk(node: Parser.SyntaxNode, type: string): Parser.SyntaxNode[] {
+  const res: Parser.SyntaxNode[] = [];
+  const stack = [node];
+  while (stack.length) {
+    const n = stack.pop();
+    if (!n) continue;
+    if (n.type === type) res.push(n);
+    stack.push(...n.namedChildren);
+  }
+  return res;
+}
+export default walk;
